requirements.txt
# Core ML / Vision
ultralytics>=8.3.0
coremltools>=8.0
numpy>=1.26.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0

# OCR
pyobjc-framework-Vision>=10.0; sys_platform == "darwin"
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
easyocr>=1.7.0

# LLM
anthropic>=0.40.0
ollama>=0.4.0

# Web dashboard
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0

# Config / validation
pydantic>=2.9.0
pydantic-settings>=2.6.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Knowledge base / matching
rapidfuzz>=3.10.0

# CLI
click>=8.1.0

# Utilities
adb-shell>=0.4.0
tqdm>=4.66.0
rich>=13.0.0
